fetchdarcs: add mirror support

+++ pkgs/build-support/fetchdarcs/builder.sh
@@ -14,8 +14,19 @@ elif test -n "$context"; then
     tagflags="--context=$context"
 fi
 
-echo "Cloning $url ${tagtext} into $out"
+# Repository list may contain ?. No glob expansion for that.
+set -o noglob
 
-darcs clone --lazy $tagflags "$url" "$out"
-# remove metadata, because it can change
-rm -rf "$out/_darcs"
+for repository in $repositories; do
+    echo "Trying to clone $repository $tagtext into $out …"
+    if darcs clone --lazy $tagflags "$repository" "$out"; then
+        # remove metadata, because it can change
+        rm -rf "$out/_darcs"
+        exit 0
+    fi
+done
+
+set +o noglob
+
+echo "Error: couldn’t clone repository from any mirror" 1>&2
+exit 1
